+++ CHANGELOG.md
@@ -14,6 +14,7 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ### Added
 * Support for decomposing projective transforms
+* Stack images now can do automatic casting between uint255 and float01
 
 
 ### Changed